* moved "latest" code to a separate Latest/ plugin folder. * remove/disabled redundant debug output logging
