+++ api/v1/Warning.php
@@ -214,45 +214,4 @@ class Warning extends Base
             );
         }
     }
-
-    /**
-     * @throws Exception
-     */
-    public function latest(float $time): PromiseInterface
-    {
-        $toPromiseFunctions[] = tpf(function () use ($time) {
-            $qb = $this->getAsyncDatabase()->createQueryBuilder();
-            return $this->getAsyncDatabase()->query(
-                $qb
-                    ->select(
-                        'warning_id as issue_database_id',
-                        'warning_active as active',
-                        'warning_layer_id as plan_layer_id',
-                        'warning_issue_type as type',
-                        'warning_x as x',
-                        'warning_y as y',
-                        'warning_source_plan_id as source_plan_id',
-                        'warning_restriction_id as restriction_id'
-                    )
-                    ->from('warning')
-                    ->where('warning_last_update > ' . $qb->createPositionalParameter($time))
-            );
-        });
-        $toPromiseFunctions[] = tpf(function () use ($time) {
-            $qb = $this->getAsyncDatabase()->createQueryBuilder();
-            return $this->getAsyncDatabase()->query(
-                $qb
-                    ->select(
-                        'shipping_warning_id as warning_id',
-                        'shipping_warning_source_geometry_persistent_id as source_geometry_persistent_id',
-                        'shipping_warning_destination_geometry_persistent_id as destination_geometry_persistent_id',
-                        'shipping_warning_message as message',
-                        'shipping_warning_active as active'
-                    )
-                    ->from('shipping_warning')
-                    ->where('shipping_warning_lastupdate > ' . $qb->createPositionalParameter($time))
-            );
-        });
-        return parallel($toPromiseFunctions);
-    }
 }
