+++ src/Domain/WsServer/Plugins/Latest/GameLatest.php
@@ -1,0 +1,485 @@
+<?php
+
+namespace App\Domain\WsServer\Plugins\Latest;
+
+use App\Domain\Common\CommonBase;
+use Drift\DBAL\Result;
+use Exception;
+use React\Promise\Deferred;
+use React\Promise\PromiseInterface;
+use function App\parallel;
+use function App\tpf;
+
+class GameLatest extends CommonBase
+{
+    /**
+     * Gets the latest plans & messages from the server
+     *
+     * @param int $teamId
+     * @param float $lastUpdateTime
+     * @param int $user
+     * @return PromiseInterface
+     * @throws Exception
+     */
+    // phpcs:ignore PSR1.Methods.CamelCapsMethodName.NotCamelCaps
+    public function Latest(int $teamId, float $lastUpdateTime, int $user): PromiseInterface
+    {
+        return $this->calculateUpdatedTime(
+        )
+        ->then(function (array $tick) use ($teamId, $lastUpdateTime, $user) {
+            $newTime = microtime(true);
+            $data = array();
+            $data['prev_update_time'] = $lastUpdateTime;
+            return $this->latestLevel2(
+                $tick,
+                $lastUpdateTime,
+                $newTime,
+                $data
+            )
+            ->then(function (array $planData) use (
+                $teamId,
+                $lastUpdateTime,
+                $user,
+                $newTime,
+                &$data
+            ) {
+                return $this->latestLevel3(
+                    $planData,
+                    $lastUpdateTime,
+                    $newTime,
+                    $data
+                )
+                ->then(function (array $layersContainer) use (
+                    $teamId,
+                    $lastUpdateTime,
+                    $user,
+                    $newTime,
+                    &$data
+                ) {
+                    return $this->latestLevel4(
+                        $layersContainer,
+                        $teamId,
+                        $lastUpdateTime,
+                        $newTime,
+                        $data
+                    )
+                    ->then(function (array $result) use (
+                        $teamId,
+                        $lastUpdateTime,
+                        $user,
+                        $newTime,
+                        &$data
+                    ) {
+                        return $this->latestLevel5(
+                            $result,
+                            $lastUpdateTime,
+                            $newTime,
+                            $data
+                        )
+                        ->then(function (Result $result) use (
+                            $teamId,
+                            $lastUpdateTime,
+                            $user,
+                            $newTime,
+                            &$data
+                        ) {
+                            return $this->latestLevel6(
+                                $result,
+                                $lastUpdateTime,
+                                $newTime,
+                                $data
+                            )
+                            ->then(function (array $results) use (
+                                $teamId,
+                                $lastUpdateTime,
+                                $user,
+                                $newTime,
+                                &$data
+                            ) {
+                                return $this->latestLevel7(
+                                    $results,
+                                    $teamId,
+                                    $lastUpdateTime,
+                                    $newTime,
+                                    $data
+                                )
+                                ->then(function (array $results) use (
+                                    $lastUpdateTime,
+                                    $user,
+                                    $newTime,
+                                    &$data
+                                ) {
+                                    return $this->latestLevel8(
+                                        $results,
+                                        $lastUpdateTime,
+                                        $newTime,
+                                        $data
+                                    )
+                                    ->then(function (array $results) use (
+                                        $lastUpdateTime,
+                                        $user,
+                                        $newTime,
+                                        &$data
+                                    ) {
+                                        return $this->latestLevel9(
+                                            $results,
+                                            $lastUpdateTime,
+                                            $newTime,
+                                            $data
+                                        )
+                                        ->then(function (Result $result) use (
+                                            $lastUpdateTime,
+                                            $user,
+                                            $newTime,
+                                            &$data
+                                        ) {
+                                            return $this->latestLevel10($result, $user, $newTime, $data)
+                                                ->then(function (/*?Result $result */) use ($data) {
+                                                    return $data;
+                                                });
+                                        });
+                                    });
+                                });
+                            });
+                        });
+                    });
+                });
+            });
+        });
+    }
+
+    /**
+     * @throws Exception
+     * @noinspection PhpSameParameterValueInspection
+     */
+    private function calculateUpdatedTime(bool $showDebug = false): PromiseInterface
+    {
+        return $this->getAsyncDatabase()->query(
+            $this->getAsyncDatabase()->createQueryBuilder()
+                ->select(
+                    'game_state as state',
+                    'game_lastupdate as lastupdate',
+                    'game_currentmonth as month',
+                    'game_start as start',
+                    'game_planning_gametime as era_gametime',
+                    'game_planning_realtime as era_realtime',
+                    'game_planning_era_realtime as planning_era_realtime',
+                    'game_planning_monthsdone as era_monthsdone',
+                    'game_mel_lastmonth as mel_lastmonth',
+                    'game_cel_lastmonth as cel_lastmonth',
+                    'game_sel_lastmonth as sel_lastmonth',
+                    'game_eratime as era_time'
+                )
+                ->from('game')
+                ->setMaxResults(1)
+        )
+        ->then(function (Result $result) use ($showDebug) {
+            $assureGameLatestUpdate = new Deferred();
+            $tick = $result->fetchFirstRow();
+            //only update if the game is playing
+            if (!in_array($tick['state'], ['END', 'PAUSE', 'SETUP']) && $tick['lastupdate'] == 0) {
+                //if the last update was at time 0, this is the very first tick happening for this game
+                $qb = $this->getAsyncDatabase()->createQueryBuilder();
+                $this->getAsyncDatabase()->query(
+                    $qb
+                        ->update('game')
+                        ->set('game_lastupdate', $qb->createPositionalParameter(microtime(true)))
+                )
+                ->done(
+                    function (Result $result) use (&$tick, $assureGameLatestUpdate) {
+                        $tick['lastupdate'] = microtime(true);
+                        $assureGameLatestUpdate->resolve();
+                    },
+                    function ($reason) use ($assureGameLatestUpdate) {
+                        $assureGameLatestUpdate->reject($reason);
+                    }
+                );
+            } else {
+                $assureGameLatestUpdate->resolve();
+            }
+            return $assureGameLatestUpdate->promise()
+                ->then(function () use ($tick, $showDebug) {
+                    $state = $tick["state"];
+                    $secondsPerMonth = $tick['era_realtime'] / $tick['era_gametime'];
+
+                    //only update if the game is playing
+                    if ($state != "END" && $state != "PAUSE" && $state != "SETUP") {
+                        $currentTime = microtime(true);
+                        $lastUpdate = $tick['lastupdate'];
+                        assert($lastUpdate != 0);
+
+                        $diff = $currentTime - $lastUpdate;
+                        $secondsPerMonth = $tick['era_realtime'] / $tick['era_gametime'];
+
+                        if ($diff < $secondsPerMonth) {
+                            $tick['era_timeleft'] = $tick['era_realtime'] - $diff -
+                                ($tick['era_monthsdone'] * $secondsPerMonth);
+                        } else {
+                            $tick['era_timeleft'] = -1;
+                        }
+
+                        if ($showDebug) {
+                            wdo("diff: " . $diff);
+                        }
+
+                        if ($showDebug) {
+                            wdo("timeleft: " . $tick['era_timeleft']);
+                        }
+                    } elseif ($state == "PAUSE" || $state == "SETUP") {
+                        //[MSP-1116] Seems sensible?
+                        $tick['era_timeleft'] = $tick['era_realtime'] - ($tick['era_monthsdone'] * $secondsPerMonth);
+                        if ($showDebug) {
+                            echo "GAME PAUSED";
+                        }
+                    } else {
+                        if ($showDebug) {
+                            echo "GAME ENDED";
+                        }
+                    }
+
+                    if ($showDebug) {
+                        wdo($tick);
+                    }
+
+                    return $tick;
+                });
+        });
+    }
+
+    /**
+     * @throws Exception
+     */
+    private function latestLevel2(array $tick, float $lastUpdateTime, float $newTime, array &$data): PromiseInterface
+    {
+        $data['tick'] = $tick;
+        if (defined('DEBUG_PREF_TIMING') && DEBUG_PREF_TIMING === true) {
+            echo (microtime(true) - $newTime) . " elapsed after tick<br />";
+        }
+        $plan = new PlanLatest();
+        $this->asyncDataTransferTo($plan);
+        return $plan->latest((int)$lastUpdateTime);
+    }
+
+    /**
+     * @throws Exception
+     */
+    private function latestLevel3(
+        array $planData,
+        float $lastUpdateTime,
+        float $newTime,
+        array &$data
+    ): PromiseInterface {
+        $data['plan'] = $planData;
+        if (defined('DEBUG_PREF_TIMING') && DEBUG_PREF_TIMING === true) {
+            echo (microtime(true) - $newTime) . " elapsed after plan<br />";
+        }
+
+        $layer = new LayerLatest();
+        $this->asyncDataTransferTo($layer);
+        $toPromiseFunctions = [];
+        foreach ($data['plan'] as $p) {
+            $toPromiseFunctions[] = tpf(function () use ($layer, $p, $lastUpdateTime) {
+                return $layer->latest($p['layers'], $lastUpdateTime, $p['id']);
+            });
+        }
+        return parallel(
+            $toPromiseFunctions
+        );
+    }
+
+    /**
+     * @throws Exception
+     */
+    private function latestLevel4(
+        array $layersContainer,
+        int $teamId,
+        float $lastUpdateTime,
+        float $newTime,
+        array &$data
+    ): PromiseInterface {
+        foreach ($data['plan'] as $key => &$p) {
+            //only send the geometry when it's required
+            $p['layers'] = $layersContainer[$key];
+            if (defined('DEBUG_PREF_TIMING') && DEBUG_PREF_TIMING === true) {
+                echo (microtime(true) - $newTime) . " elapsed after layers<br />";
+            }
+
+            if ((
+                $p['state'] == "DESIGN" && $p['previousstate'] == "CONSULTATION" &&
+                $p['country'] != $teamId
+            )) {
+                $p['active'] = 0;
+            }
+        }
+        unset($p);
+
+        $plan = new PlanLatest();
+        $this->asyncDataTransferTo($plan);
+        return $plan->getMessages(
+            $lastUpdateTime
+        );
+    }
+
+    /**
+     * @throws Exception
+     */
+    private function latestLevel5(
+        array $queryResultRows,
+        float $lastUpdateTime,
+        float $newTime,
+        array &$data
+    ): PromiseInterface {
+        $data['planmessages'] = $queryResultRows;
+        if (defined('DEBUG_PREF_TIMING') && DEBUG_PREF_TIMING === true) {
+            echo (microtime(true) - $newTime) . " elapsed after plan messages<br />";
+        }
+
+        //return any raster layers that need to be updated
+        $layer = new LayerLatest();
+        $this->asyncDataTransferTo($layer);
+        return $layer->latestRaster(
+            $lastUpdateTime
+        );
+    }
+
+    /**
+     * @throws Exception
+     */
+    private function latestLevel6(
+        Result $queryResult,
+        float $lastUpdateTime,
+        float $newTime,
+        array &$data
+    ): PromiseInterface {
+        $data['raster'] = $queryResult->fetchAllRows();
+        if (defined('DEBUG_PREF_TIMING') && DEBUG_PREF_TIMING === true) {
+            echo (microtime(true) - $newTime) . " elapsed after raster<br />";
+        }
+
+        $energy = new EnergyLatest();
+        $this->asyncDataTransferTo($energy);
+        return $energy->latest(
+            $lastUpdateTime
+        );
+    }
+
+    /**
+     * @param Result[] $queryResults
+     * @param float $lastUpdateTime
+     * @param int $teamId
+     * @param float $newTime
+     * @param array $data
+     * @return PromiseInterface
+     * @throws Exception
+     */
+    private function latestLevel7(
+        array $queryResults,
+        int $teamId,
+        float $lastUpdateTime,
+        float $newTime,
+        array &$data
+    ): PromiseInterface {
+        $data['energy']['connections'] = $queryResults[0]->fetchAllRows();
+        $data['energy']['output'] = $queryResults[1]->fetchAllRows();
+
+        if (defined('DEBUG_PREF_TIMING') && DEBUG_PREF_TIMING === true) {
+            echo (microtime(true) - $newTime) . " elapsed after energy<br />";
+        }
+
+        $kpi = new KpiLatest();
+        $this->asyncDataTransferTo($kpi);
+        return $kpi->latest(
+            (int)$lastUpdateTime,
+            $teamId
+        );
+    }
+
+    /**
+     * @throws Exception
+     */
+    private function latestLevel8(
+        array $queryResultRows,
+        float $lastUpdateTime,
+        float $newTime,
+        array &$data
+    ): PromiseInterface {
+        $data['kpi'] = $queryResultRows;
+
+        if (defined('DEBUG_PREF_TIMING') && DEBUG_PREF_TIMING === true) {
+            echo (microtime(true) - $newTime) . " elapsed after kpi<br />";
+        }
+
+        $warning = new WarningLatest();
+        $this->asyncDataTransferTo($warning);
+        return $warning->latest(
+            $lastUpdateTime
+        );
+    }
+
+    /**
+     * @throws Exception
+     */
+    private function latestLevel9(
+        array $queryResults,
+        float $lastUpdateTime,
+        float $newTime,
+        array &$data
+    ): PromiseInterface {
+        $data['warning']['plan_issues'] = $queryResults[0]->fetchAllRows();
+        $data['warning']['shipping_issues'] = $queryResults[1]->fetchAllRows();
+
+        if (defined('DEBUG_PREF_TIMING') && DEBUG_PREF_TIMING === true) {
+            echo (microtime(true) - $newTime) . " elapsed after warning<br />";
+        }
+
+        $objective = new ObjectiveLatest();
+        $this->asyncDataTransferTo($objective);
+        return $objective->latest(
+            $lastUpdateTime
+        );
+    }
+
+    /**
+     * @throws Exception
+     */
+    private function latestLevel10(
+        Result $result,
+        int $user,
+        float $newTime,
+        array &$data
+    ): ?PromiseInterface {
+        $data['objectives'] = $result->fetchAllRows();
+
+        if (defined('DEBUG_PREF_TIMING') && DEBUG_PREF_TIMING === true) {
+            echo (microtime(true) - $newTime) . " elapsed after objective<br />";
+        }
+
+        //Add a slight fudge of 1ms to the update times to avoid rounding issues.
+        $data['update_time'] = $newTime - 0.001;
+
+        // send an empty string if nothing was updated
+        if (empty($data['energy']['connections']) &&
+            empty($data['energy']['output']) &&
+            empty($data['geometry']) &&
+            empty($data['plan']) &&
+            empty($data['messages']) &&
+            empty($data['planmessages']) &&
+            empty($data['kpi']) &&
+            empty($data['warning']) &&
+            empty($data['raster']) &&
+            empty($data['objectives'])) {
+            $data = '';
+            return null;
+        }
+
+        return $this->getAsyncDatabase()->update(
+            'user',
+            [
+                'user_id' => $user
+            ],
+            [
+                'user_lastupdate' => $newTime
+            ]
+        );
+    }
+}
